refactor: extract mode and inspection policy leaves
