+++ scripts/build_antigravity_distribution.py
@@ -28,7 +28,9 @@ HOOK_FILES = (
     "click_host_router.py",
     "click_import_bootstrap.py",
     "click_inspection.py",
+    "click_inspection_policy.py",
     "click_lifecycle.py",
+    "click_mode.py",
     "click_mutation.py",
     "click_observation.py",
     "click_process.py",
