+++ tests/repository_policy_core.py
@@ -535,7 +535,9 @@ class RepositoryPolicyTests(unittest.TestCase):
                 "click_host_router.py",
                 "click_import_bootstrap.py",
                 "click_inspection.py",
+                "click_inspection_policy.py",
                 "click_lifecycle.py",
+                "click_mode.py",
                 "click_gate.py",
                 "click_mutation.py",
                 "click_observation.py",
